+++ CHANGELOG.md
@@ -5,7 +5,15 @@ All notable changes to this project will be documented in this file. This projec
 
 ## Unreleased
 
-## [3.2.0] - 2024-04-19
+## [3.3.0] - 2025-05-24
+
+### Added
+
+- New `enum_value` and `enum_string` helper functions for getting a scalar or string value from an enum.
+- The `Guid::fromUuid()` method now also accepts a `Uuid` object, so that it can be used to create a GUID from a UUID
+  identifier. Previously it required a Ramsey UUID object.
+
+## [3.2.0] - 2025-04-19
 
 ### Added
 
@@ -415,6 +423,8 @@ All notable changes to this project will be documented in this file. This projec
 ## [0.1.0] - 2023-11-18
 
 Initial release.
+
+[3.3.0]: https://github.com/cloudcreativity/ddd-modules/compare/v3.2.0...v3.3.0
 
 [3.2.0]: https://github.com/cloudcreativity/ddd-modules/compare/v3.1.0...v3.2.0